Refactored main window ui into new file.
